requirements.txt
Flask==2.3.3
pandas==2.2.1
numpy>=1.24.0,<2.0
scikit-learn==1.4.2
joblib==1.3.2
mysql-connector-python==8.3.0
shap==0.44.1
matplotlib==3.8.4
gunicorn==21.2.0
